fix: Move const variables to repair model
